Add searching and float speed
Finished basic search feature where chaser AI only chases you if you are within a certain radius, and changed Mob.move() to allow for floating-point speed values. Also required changing positions to doubles and casting to ints when rendering.

+++ src/com/davidflyr/radtodd/entity/Entity.java
@@ -7,7 +7,7 @@ import com.davidflyr.radtodd.level.Level;
 
 public class Entity {
 	
-	protected int x, y;
+	protected double x, y;
 	private boolean removed = false;
 	protected Level level;
 	protected final Random random = new Random();
@@ -20,11 +20,11 @@ public class Entity {
 		
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return y;
 	}
